% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_data.R
\name{cr_data-accessors}
\alias{cr_data-accessors}
\alias{cr_get_data}
\alias{cr_metadata}
\title{Accessor methods for \code{cr_data} objects}
\usage{
cr_get_data(x)

cr_metadata(x)
}
\arguments{
\item{x}{A \code{cr_data} object.}
}
\value{
See individual descriptions above.
}
\description{
\describe{
  \item{\code{cr_get_data}}{Returns the data frame stored in the
    \code{data} slot.}
  \item{\code{cr_metadata}}{Returns a named list of all metadata slots
    (\code{causes}, \code{var_names}, \code{var_types},
    \code{time_var}, \code{event_var}, \code{id_var},
    \code{sort_by_time}, \code{time_offset}, \code{cens_code}).}
}
}
