% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cr_models.R
\name{list_cr_models}
\alias{list_cr_models}
\title{List all registered model keys}
\usage{
list_cr_models()
}
\value{
A character vector of registered model keys (may be empty).
}
\description{
List all registered model keys
}
\examples{
list_cr_models()
}
