% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif_io.R
\name{pool_cifs_mean}
\alias{pool_cifs_mean}
\title{Pool a list of CIF arrays by taking the element-wise mean}
\usage{
pool_cifs_mean(cif_list)
}
\arguments{
\item{cif_list}{A list of 3-D numeric arrays, each of dimension
`[n, K, Tm]`.}
}
\value{
A single 3-D array of the same dimensions.
}
\description{
Used to combine predictions across multiple imputed datasets.
}
