\name{list_cr_models}
\alias{list_cr_models}
\alias{list_models}
\title{List registered model keys}
\usage{
list_cr_models()

list_models()
}
\description{
Returns a character vector of all model keys currently in the registry.
\code{list_models()} is an alias for \code{list_cr_models()}.
}
\value{
A character vector of registered model keys.
}
\seealso{
\code{\link{register_cr_model}}
}
