% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trapezoidal.integration}
\alias{trapezoidal.integration}
\title{Trapezoidal numerical integration}
\usage{
trapezoidal.integration(x, y)
}
\arguments{
\item{x}{Numeric vector of x-values (must be sorted).}

\item{y}{Numeric vector of y-values (same length as `x`).}
}
\value{
A single numeric value.
}
\description{
Trapezoidal numerical integration
}
